Reorganize all k8s module with new chaos functions
