+++ node/tests/src/utils.rs
@@ -17,7 +17,7 @@ pub(crate) async fn deploy_role() -> anyhow::Result<()> {
     let client = k8s::get_client().await?;
     k8s::create_or_reuse_namespace(&client, k8s::DEFAULT_NAMESPACE).await?;
     k8s::create_or_reuse_pod_reader_role(&client).await?;
-    k8s::create_or_reuse_network_chaos_role(&client).await?;
+    k8s::chaos::create_or_reuse_network_chaos_role(&client).await?;
     Ok(())
 }
