+++ node/tools/src/k8s/mod.rs
@@ -1,0 +1,4 @@
+pub mod chaos;
+mod operations;
+
+pub use operations::*;
