+++ node/tools/src/lib.rs
@@ -2,7 +2,6 @@
 #![allow(missing_docs)]
 mod config;
 pub mod k8s;
-pub mod network_chaos;
 mod proto;
 pub mod rpc;
 mod store;
